+++ Modules/DiffusionImaging/FiberTracking/Algorithms/TrackingHandlers/mitkTrackingHandlerRandomForest.h
@@ -78,15 +78,15 @@ public:
     void SetNumPreviousDirections( int num ){ m_NumPreviousDirections=num; }
     void SetNumTrees(int num){ m_NumTrees = num; }
     void SetMaxTreeDepth(int depth){ m_MaxTreeDepth = depth; }
-    void SetStepSize(double step){ m_WmSampleDistance = step; }
+    void SetStepSize(float step){ m_WmSampleDistance = step; }
     void SetGrayMatterSamplesPerVoxel(int samples){ m_GmSamplesPerVoxel = samples; }
-    void SetSampleFraction(double fraction){ m_SampleFraction = fraction; }
+    void SetSampleFraction(float fraction){ m_SampleFraction = fraction; }
     void SetBidirectionalFiberSampling(bool val) { m_BidirectionalFiberSampling = val; }
     void SetZeroDirWmFeatures(bool val) { m_ZeroDirWmFeatures = val; }
     std::shared_ptr< vigra::RandomForest<int> > GetForest(){ return m_Forest; }
 
     void InitForTracking();     ///< calls InputDataValidForTracking() and creates feature images
-    vnl_vector_fixed<double,3> ProposeDirection(itk::Point<double, 3>& pos, std::deque< vnl_vector_fixed<double,3> >& olddirs, itk::Index<3>& oldIndex);  ///< predicts next progression direction at the given position
+    vnl_vector_fixed<float,3> ProposeDirection(itk::Point<float, 3>& pos, std::deque< vnl_vector_fixed<float,3> >& olddirs, itk::Index<3>& oldIndex);  ///< predicts next progression direction at the given position
 
     bool IsForestValid();   ///< true is forest is not null, has more than 0 trees and the correct number of features (NumberOfSignalFeatures + 3)
 
@@ -126,20 +126,20 @@ protected:
     std::vector< ItkUcharImgType::Pointer >                     m_MaskImages;               ///< binary mask images to constrain training to a certain area (e.g. brain mask)
     std::vector< ItkUcharImgType::Pointer >                     m_WhiteMatterImages;        ///< defines white matter voxels. if not set, theses mask images are automatically generated from the input tractograms
 
-    double                                                      m_WmSampleDistance;         ///< deterines the number of white matter samples (distance of sampling points on each fiber).
+    float                                                       m_WmSampleDistance;         ///< deterines the number of white matter samples (distance of sampling points on each fiber).
     int                                                         m_NumTrees;                 ///< number of trees in random forest
     int                                                         m_MaxTreeDepth;             ///< limits the tree depth
-    double                                                      m_SampleFraction;           ///< fraction of samples used to train each tree
+    float                                                       m_SampleFraction;           ///< fraction of samples used to train each tree
     unsigned int                                                m_NumberOfSamples;          ///< stores overall number of samples used for training
     std::vector< unsigned int >                                 m_GmSamples;                ///< number of gray matter samples
     int                                                         m_GmSamplesPerVoxel;        ///< number of gray matter samplees per voxel. if -1, then the number is automatically chosen to gain an overall number of GM samples close to the number of WM samples.
-    vigra::MultiArray<2, double>                                m_FeatureData;              ///< vigra container for training features
+    vigra::MultiArray<2, float>                                 m_FeatureData;              ///< vigra container for training features
     unsigned int                                                m_NumPreviousDirections;        ///< How many "old" directions should be used as classification features?
 
     // only for tracking
-    vigra::MultiArray<2, double>                                m_LabelData;                    ///< vigra container for training labels
+    vigra::MultiArray<2, float>                                 m_LabelData;                    ///< vigra container for training labels
     vigra::MultiArray<2, double>                                m_Weights;                      ///< vigra container for training sample weights
-    std::vector< vnl_vector_fixed<double,3> >                   m_DirectionContainer;
+    std::vector< vnl_vector_fixed<float,3> >                    m_DirectionContainer;
 
     bool                                                        m_BidirectionalFiberSampling;
     bool                                                        m_ZeroDirWmFeatures;
